Translate faq-versioning to spanish
There's a few words that I am not 100% sure about the translation, some of those are:
* breaking changes
* major/minor version
* package

I tried to be consistent as much as possible. I also changed the link that pointed to the english version of the semver docs to [the one in spanish](https://semver.org/lang/es/).